obs-video-scheduler/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
pydantic==2.7.4
pydantic-settings==2.2.1
obs-websocket-py==1.1.0
